Move variant inference methods to own file
